Revert "This is a massive round of changes to make the cross-compile 64-bit"
This reverts commit 3e4fc5ac8ca502fbc8d7465881cc37448262ad44.

Turns out there's a bajillion places where pointers are included in
structs and then the whole thing is written out to disk.  Unfortunately,
anyone wanting to build this thing is just gonna have to have the
ability to build 32-bit binaries.

+++ include/artifact.h
@@ -39,15 +39,15 @@
 struct artifact {
 	short	    otyp;
 	const char  *name;
-	uint32_t spfx;	/* special effect from wielding/wearing */
-	uint32_t cspfx;	/* special effect just from carrying obj */
-	uint32_t mtype;	/* monster type, symbol, or flag */
+	unsigned long spfx;	/* special effect from wielding/wearing */
+	unsigned long cspfx;	/* special effect just from carrying obj */
+	unsigned long mtype;	/* monster type, symbol, or flag */
 	struct attack attk, defn, cary;
 	uchar	    inv_prop;	/* property obtained by invoking artifact */
 	aligntyp    alignment;	/* alignment of bequeathing gods */
 	short	    role;	/* character role associated with */
 	short	    race;	/* character race associated with */
-	int32_t        cost;	/* price when sold to hero (default 100 x base cost) */
+	long        cost;	/* price when sold to hero (default 100 x base cost) */
 };
 
 /* invoked properties with special powers */
