+++ include/prop.h
@@ -79,7 +79,7 @@
 /* Definitions were moved here from obj.h and you.h */
 struct prop {
 	/*** Properties conveyed by objects ***/
-	int32_t extrinsic;
+	long extrinsic;
 	/* Armor */
 #	define W_ARM	    0x00000001L /* Body armor */
 #	define W_ARMC	    0x00000002L /* Cloak */
@@ -112,10 +112,10 @@ struct prop {
 #	define W_CHAIN	    0x00400000L /* Punishment chain */
 
 	/*** Property is blocked by an object ***/
-	int32_t blocked;					/* Same assignments as extrinsic */
+	long blocked;					/* Same assignments as extrinsic */
 
 	/*** Timeouts, permanent properties, and other flags ***/
-	int32_t intrinsic;
+	long intrinsic;
 	/* Timed properties */
 #	define TIMEOUT	    0x00ffffffL /* Up to 16 million turns */
 	/* Permanent properties */
