+++ src/do.c
@@ -640,7 +640,7 @@ menu_drop(retry)
 int retry;
 {
     int n, i, n_dropped = 0;
-    int32_t cnt;
+    long cnt;
     struct obj *otmp, *otmp2;
 #ifndef GOLDOBJ
     struct obj *u_gold = 0;
@@ -1574,7 +1574,7 @@ struct obj *corpse;
 void
 revive_mon(arg, timeout)
 genericptr_t arg;
-int32_t timeout;
+long timeout;
 {
     struct obj *body = (struct obj *) arg;
 
@@ -1635,7 +1635,7 @@ dowipe()
 
 void
 set_wounded_legs(side, timex)
-register int32_t side;
+register long side;
 register int timex;
 {
 	/* KMH -- STEED
